correcting references in downloader test. Adding global fixture to mock redis. Altering manager class to mock DB manager to avoid threading issues
